stage 1/requirements.txt
pip>=24.0
python-dotenv>=1.0.1
